+++ asu/utils/boss.py
@@ -3,31 +3,33 @@ from queue import Queue
 import logging
 import time
 
-from utils.image import Image
-from utils.common import get_hash
-from utils.config import Config
-from utils.database import Database
-from utils.worker import Worker
+from asu.utils.image import Image
+from asu.utils.common import get_hash
+from asu.utils.config import Config
+from asu.utils.database import Database
+from asu.utils.worker import Worker
 
 class Boss(threading.Thread):
     def __init__(self):
         threading.Thread.__init__(self)
-        self.log = logging.getLogger("boss")
+        self.log = logging.getLogger(__name__)
         self.config = Config()
         self.database = Database(self.config)
         self.build_queue = Queue(1)
 
     def run(self):
         workers = []
-        for worker_location in self.config.get("workers"):
+        for worker_location in self.config.get("worker", []):
             worker = Worker(worker_location, "image", self.build_queue)
             worker.start()
             workers.append(worker)
 
+        self.log.info("Active workers are %s", workers)
+
         while True:
             build_job = self.database.get_build_job()
             if build_job:
-                self.log.info("found build job %s", build_job)
+                self.log.info("Found build job %s", build_job)
                 self.build_queue.put(build_job)
             else:
-                time.sleep(5)
+                time.sleep(10)
